requirements.txt
pytest
mysql-connector